Extract view for blank form list items
